edicraft-agent/requirements.txt
bedrock-agentcore
strands-agents
strands-agents-tools
rcon
requests
pytest
boto3
botocore
